move namespace scanners from polymer to javascript
